Delete asynchronously to handle long response times
When the user presses the delete button, immediately delete the todo
without waiting for a response from the server. This entails that when
the response time is high, the user won't be tempted to press the button
multiple times. Multiple presses could lead to them accidentally
deleting the todo that moves into the position of the just deleted todo.