Rewrite front end
This commit rewrites the front end. One change of note is the use of a
"task queue" to ensure that logout occurs only after todos have finished
syncing.

This commit  also changes the appendTodo API to move todo ID creation to
the client so that todos can be appended optimistically.